datanode/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
python-decouple
loguru==0.7.2
protobuf==4.25.2
